Correct invalid import paths in `index.d.ts` files.
Prior to this change, the `index.d.ts` files were moved up one
directory (and named for their package) but the import paths were not
being updated so that the resulting ES6 or AMD output referenced invalid
paths.

This change adds define.alias shims for each of the packages, and leaves
the `index.d.ts` files within the same level as they are on disk
(therefore keeping the import paths correct).

In Ember we will need a little more work during integration
(specifically the Ember loader will need to either gain aliases or
support for `index` fallback).