[WIP] Unify appends
There is no need to create a specific object for normalizing, inserting
and updating a value. Instead we should be able to grab a singleton
manager from the VM (currently this seems like overkill) that knows how
to deal with values. We should be able to rationalize this even further
and create a "singleton manager" that knows how to insert the ambigous
PathExpression that evaluates to a component definition.