Ensure custom element fallback properly specializes statements.
The root cause of the various `implementation missing for: x` errors
is that the custom element fallback was using unspecialized opcodes
that are not used/tested through the rest of the system (since we
generally use the specialized versions if at all possible).

This fixes the issue by specializing inside the fallback branch.

A joint @krisselden / @rwjblue production.