Part 1: Make most of the constants serializable
This is the initial pass of making the program serializable. In general,
previous work has made the program itself serializable, because the
program itself is an array of opcodes, and an opcode is an array of
numbers (128 bits).

However, the program often has references into the constant pool. Some
of the constant pool is already serializable (for example, strings).
However, before this series of commits, the constant pool also contained
a number of direct references to live objects that could not be
trivially serialized.

The purpose of these commits is to make all constants serializable.

To make more objects supplied by the user serializable, we introduced a
Resolver (similar but not identical to the interface in @glimmer/di)
to indirect any lookup of user objects. The lookup specifier is required
to be serializable.

As part of this work, we added a few new features, because certain
important functionality was previously implemented in userland, and the
API for implementing them was excessively dynamic.

- Implemented `{{component}}` and `(component)` (from Glimmer.js and
  Ember) natively in glimmer-vm.
- Support for first-class argument currying in `(component)`.