Part 2: All internals now support serialization
Previously, blocks were always compiled "in-place", taking advantage of
the fact that nested blocks were never compiled during the compilation
of their parent.

Now, eager compilation requires us to compile nested blocks ahead of
time, so the naive strategy of compiling into the "current" position
doesn't work anymore. Instead, we compile into a temporary buffer, and
commit into the main program once a particular block is finished. In
practice, this means that nested blocks always appear *before* their
parent in the program.

Also added initial support for a component capability mask, which
allowed us to build static component invocation. The most important
consequence is that the system is no longer architected around the quirk
of Ember's late-bound layouts. If a component definition is able to
specify its layout ahead of time, we can generate a specialized
invocation for the component we're invoking, rather than a bunch of
dynamic checks that require us to keep static information around at
runtime.

Similarly, instead of modelling dynamic blocks as a special kind of
runtime object, we now push the symbol table and Handle onto the stack.
Because the symbol table is serializable (and Handle is an integer),
this means that the constants used by dynamic invocation are always,
themselves, serializable.