[WIP]
Initial extraction of the opcode compiler into its own package.

The bulk of this work is trimming down incidental dependencies that the
opcode compiler has on other runtime types. For the most part, they are
either completely unnecessary or can be easily dependency injected.

The next step is to get `runtime` to use this package.

The step after that is to get `bundle-compiler` to use this package to
produce a fully precompiled output.

The most notable change in the interfaces in this commit is that the
`compilation options`, which are now expected to be used in the eager
bundler as well, no longer include the full resolver. Instead, they
include a `getCapabilities` method that requires only the capabilities
for a given component, and not access to arbitrary runtime objects.