Inject OpcodeBuilder
Prior to this commit parts of the compiler hardcoded the references to the LazyOpcodeBuilder. To be able to switch between eager and lazy compilation we need to pass an OpcodeBuilderConstructor function into all the places where we were hardcoding the LazyOpcodeBuilder, specifically CompilableTemplate and WrappedTemplate. By treating the Builder's concrete type unknown, we can swap between builders.