Allow AOT mode to invoke static directly
Lazy mode does static invocation by pushing the uncompiled block into
the constants, then following by a CompileBlock instruction, which puts
the handle on the stack, followed by an InvokeVirtual instruction, which
invokes the handle at the top of the stack.

AOT mode compiles the blocks at compile time, so they can include the
handle in the instruction itself.

This commit distinguishes between InvokeVirtual, which expects a Handle
at the top of the stack (which is also necessary to yield to blocks),
and InvokeStatic, which includes the Handle in the opcode, and is able
to avoid using the stack.