[BUGFIX] Fix SVG and Block Rehydration
This continues the integrity checks for rehydration and fixes 2 critical bugs related to SVG elements and blocks. Prior to this PR we would attempt to read the `node.tagName` off each element that was the candidate for rehydration. For HTMLElements this will return a value that is all uppercase, however for elements that are SVG the tagName will be lowercase :(. This meant the SVG would get re-created during rehydration because there was a mismatch in tagName.

The other bug was much worst when it came to cases where the client state and server state did match up which caused blocks to open or close for whatever reason. Nested blocks would not clear out and blocks that opened on the client would be inserted past the end block markers. While this didn't seem fatal, was pretty weird that the metamorphs were retained.