Store JS values in a different stack from numbers
After this change, the main stack is an array of u32s. If a u32 is
pushed onto the stack, it is treated as an immediate and stored
directly. If any other value is pushed onto the stack, it is pushed onto
a secondary array of JavaScript values, and it's stored as a pointer
(a u32 offset with the high bit set).

When getting a value from the stack, if the high bit is set, the value
is loaded from the JavaScript stack, otherwise the immediate is
returned.

The goal of this work is to allow more of the VM to operate inside of
wasm-ish code (operating on numbers and simple data structures). The
endgame is to cleanly separate between normal VM operations (jumps,
pushes, pops, dups, etc.) and specialized VM operations that work with
JS objects ("syscalls") and to try to do a lot of contiguous work inside
the VM, and batch work that requires syscalls.

There are many additional improvements I plan to make on top of this
work:

- operations like dup() currently dereference the pointer and push it
  back onto the stack. This both requires the VM to unnecessarily work
  with JavaScript objects and also is wasteful in terms of work. It
  would be straight forward to modify dup() to duplicate the pointer
  directly and not worry about the JS stack.
- A number of simple values (null, undefined, true, false) are already
  encoded in the VM as u32s in some places. Today, since they are not
  u32s, they become pointers into the JavaScript stack. We should beef
  up the tagged pointer system to support more values as immediates.
- Avoid pushing null onto the JS stack when the value is an immediate.
  This is a temporary hack so I don't need to decide the allocation
  strategy for the JS stack. Two options:
  1. since appending is a bounded operation, just use a single arena for
  the entire append process
  2. split up the fp into a u16 pointer into the u32 stack and a u16
  pointer into the js stack. This wouldn't require us to do anything in
  JS at the point where we popped the stack, it would just allow us to
  reuse slots in the JS stack throughout the appending process.

Without these improvements, there are already ~10,000 immediates on the
stack in the test suite, compared to ~30,000 pointers (in large part
because there's a decent amount of stack management around frames that
is purely numeric).