Revert some unneeded changes to `PartialInvoker`
The original fix is more or less correct, however, we shouldn't have
to change anything related to `EvalScope`, because that dictionary is
representing what was "passed in" to the component (named args and
blocks). Since a partial cannot introduce additional arguments that
part of the code is not a problem.

However, there is a separate bug that was caught by the new test. The
"top-level" template rendering helper does not bind the `EvalScope`
even when `hasPartial` is true. (This is because a "top-level" template
cannot  have args and blocks passed in.) However, the `PartialInvoker`
assumes it is always present (`!`). So a seconary fix is needed there
to correct for the problem.