Port `LowLevelVM` to WebAssembly
This commit ports a big chunk of the `LowLevelVM` implementation into
WebAssembly and Rust. This involved quite a bit of juggling of various items
here and there and the Rust code expanded relatively significantly. A number of
callbacks are now set up between Rust/wasm as not everything has been ported to
Rust yet but I think the idea is that more and more of the VM code will migrate
towards Rust over time.

As-is the tests should continue to pass and nothing should regress from the
previous commit. Next up should hopefully be even more JS to Rust!