Abstract main thread JS operations using InstructionList data structure
To avoid the performance overhead of context switching between JavaScript and WebAssembly, it is ideal to perform as much work in the current context as possible before yielding back.

Unfortunately, the current implementation of opcodes assumes direct access to JavaScript objects, including our DOM building abstractions like ElementBuilder. This makes these opcodes difficult to port to WebAssembly without having them frequently cross the JS/wasm boundary.

This commit introduces an intermediate binary data structure, called an InstructionList, for encoding a list of operations that can only be performed in the main JavaScript thread. Rather than producing side effects immediately, opcodes add an instruction to the list of instructions for the main thread to execute.
Once the thread of execution returns to the JavaScript implementation of the Glimmer VM, it executes any pending instructions before resuming execution of opcodes.

The implementation is primarily made up of two classes: the InstructionListEncoder, which efficiently records instructions into an ArrayBuffer, and the InstructionListExecutor, which iterates over that ArrayBuffer and executes each instruction in turn.

This first spike only includes a few basic DOM-related operations, but this could generalize to many more types of operations, like invoking component lifecycle hooks, etc. As more and more opcodes are moved into WebAssembly, we should be able to drop the JavaScript implementation of the InstructionListEncoder altogether, keeping just the relatively compact InstructionListExecutor for actually dispatching instructions.