Try to use the stack instead of explicit scopes
The bigger picture of this commit is that I'd like to use the stack as a
scope in all cases, so that the low-level part of the code doesn't need
to worry about interacting with a higher-level Scope object.

After a few attempts to jump to the punchline, I discovered that the
Scope entangles too many concerns, and this commit starts untangling
them a little at a time.

First, I'll try to move more cases into the StackScopeProxy, and once
that's all done, look at whether I can eliminate it altogether (and just
rely on frames and $fp to look up stack variables).