Switch virtual component calls to use stack scopes
With static calls, the compiler can set up the stack to match the layout
expected by the callee. With virtual calls, we need to copy values from
the caller's layout into the callee's layout.

At the moment, I copy the values out of the ARGS mapping into the
callee's layout, and create a stack Scope proxy whose base points at the
start of the copied values.

Eventually, I want to eliminate the need for the Scope entirely, and
just rely on the $fp, which will require some additional juggling.