[BUGFIX] Prevents caching when a cycle exists in tags
When a cycle exists between two or more tags, it's possible for the
cache to get into a bad state currently. For example, following these
steps:

1. Create two UpdatableDirtyableTags, Tag A and Tag B.
2. Update each tag to point to the other, creating a cycle.
3. Dirty Tag B
4. Get the value of Tag B, before getting the value of Tag A

When calculating the value of Tag B, it will look for the value of Tag A.
Tag A will then attempt to get the value of Tag B, will see that the
last checked revision has matches the current revision, and will return
the last known value.

This is logically consistent and prevents infinite recursion, but the
since Tag B was just updated, this value will be out of date just after
Tag A finishes calculating, by which time Tag A will unfortunately have
already cached its value against the latest revision. Tag A's cache will
remain in an inconsistent state until the next bump of the global
revision counter.

This PR places an `isUpdating` flag on CachedTags. If a cycle is
detected while updating, the last value is still returned, but we _also_
bump the global revision counter and reset the lastRevision for the tag.
This way, the tag which began the cycle will remain cached, but any tags
which were calculated _during_ the cycle will not, allowing them to
recalculate the next time they are retrieved with the updated entry
value (if it changed at all).

An unfortunate side effect of this is that any tags calculated and
cached _before_ entering the cycle will also be busted. We could
potentially avoid this by using additional bookkeeping - a set of cycle
endpoints, combined with a counter for the number of cycles seen
(potentially), but given this particular case is likely to be uncommon
in the future with autotracking (cycles can't occur if all your tags
represent leaves), I wanted to start simple.