[BUGFIX] ensure untracked `@tracked` initializers
Given...

```js
class Thing {
  @tracked foo = `foo ${this.bar}`;
  @tracked bar = 'bar';
}
```

When this is rendered...

```hbs
{{this.foo}}
```

The curly will invalidates when `this.bar` invalidates, if it
happens to be the first thing that accesses `this.foo` on the
object.

This is incorrect: initializers only run once, and whatever
code were run inside should never cause the tracked property
to become invalidated.