[PERF] Improves the autotracking transaction by creating fewer Errors
Currently we create an Error every time a value is consumed, regardless
of whether or not it is being consumed for the first time (and thus we
need the Error for its stack) or if we've already consumed it before,
and don't need an Error at all. This PR changes this so we do it only
when necessary.