[REFACTOR] Removes AoT mode
The existing AoT mode is significantly different than the future
direction that we want to go with precompiling templates. In particular,
it relied on static analysis of programs at compile time, including
being able to resolve all components ahead of time. With the direction
we're moving with template imports, this will not be possible without a
significant investment in static build tooling, so we want to focus on a
solution that does not require this.

This PR removes the existing AoT mode and refactors everything to assume
that there is only one major mode for Glimmer apps to run in, which is
effectively "JIT" mode as it exists today. We will continue to iterate
on and improve JIT mode, until it can precompile and ship byte code
templates like AoT mode could.