Upgrade to TypeScript 4.0
This also updates the eslint-typescript packages

TypeScript 4.0 invalidated the way that we were casting SimpleDOM to
browser DOM and vice versa. This commit changes our casting strategy so
that casts go through additional checks that accomodate TypeScript's new
stricter requirements.

In general, this was a good change: it adds some checks to the blind
casts we were making, and caught at least one genuine mistake in our
casts (casting a textarea to input).

However, these checks add some minor overhead and should be unnecessary
in production. This commit structures the casts so they can be stripped
out relatively easily in the future, and we should do so.