Cache `hbsPosFor(offset)` in `Source`
`hbsPosFor` is currently the hottest path in the current implementation
of the template compiler in terms of self time. Adding this cache for
it improves throughput by about half a second in one rough benchmark
(against a collection of about 1,500 templates totaling about 54,000
lines of Handlebars), run 5 times against each version of the
compiler:

|           variant            | average | min  | max  |
| ---------------------------- | ------- | ---- | ---- |
| @glimmer/compiler@0.65.3     | 7002.2  | 6976 | 7031 |
| @glimmer/compiler@master     | 9181.4  | 9159 | 9202 |
| @glimmer/compiler@experiment | 8445.2  | 8400 | 8507 |

Additionally, `hbsPosFor` *stops* being the largest self time; in the
same rough benchmark, it dropped from ~560ms to ~97ms of self time.