Use attributes for more form cases
The current tree builder uses an HTML buffer to create elements, which
means that all attributes get set atomically and before properties.

This resolves a category of bugs exemplified by
https://github.com/whatwg/html/issues/2427 *if* `type`, `min`, `max`
(and other similar attributes) are set as attributes. This means that
the element will get fully and properly set up before the `value`
property is set.

The tests still rely on the fact that `disabled="anything"` gets set as
a property, coerced into a boolean, and then reflected back to the
`disabled` attribute as `true` (resulting in `disabled=''`).

This commit attempts to resolve those issues by using properties on
input fields for:

- value
- any boolean attributes like `checked`, `disabled`, `multiple`

This means that things like `pattern`, `min` and `max` are set as
attributes, which causes the semantics of form constraints to perfectly
match HTML's semantics (because we're literally creating an element
from HTML).