Finish overhauling debug logging
The logging should be significantly more useful, and the metadata is now
extensible to support arbitrary assertions.

I already added stack type assertions to the metadata, but these are
currently still implemented in terms of the old stack change check.

The improvements to metadata already improve the logs (because they
facilitate improved deserialization of immediate values in operands).

The next step is to assert the stack types and also log the (formatted)
stack parameters and return values in the log.

All of this facilitates quicker iteration on the in-progress error
recovery PR.