infra(debugger): Significant improvement to stack checks and trace logging
This commit overhauls opcode metadata:

Previously, many opcodes opted out of stack changes because their
changes were too dynamic (e.g. Pop reduces the stack size by a
parameter, which was previously too dynamic to be checked). This commit
enables dynamic stack checks, which are functions that take the runtime
op and debug VM state and dynamically determine the stack change.

This makes it possible to check the stack for every opcode except
`EnterList`, `Iterate`, `PrepareArgs` and `PopFrame` (the reasons for
these exceptions is documented in `opcode-metadata.ts`).

Dynamic checking is now used in:

- Concat (pops N from the stack based on an operand)
- Pop (same)
- BindDynamicScope (binds a number of names from an operand)

A previous commit enabled operand metadata. That infrastructure allows
the trace logger and compilation logger to print friendly versions of
opcodes.

This commit makes the metadata pervasively more accurate, piggy-backing
on the previous commit, which made nullable operands more accurate.

This deserialization process serves as a sort-of verification pass. If
an opcode is expecting an array, for example, and the operand
deserializes to a string, it will fail.

It currently fails confusingly (and not reliably) when the deserializer
fails to deserialize, but this can and should be improved in the future.

Enabling pervasive stack checks caused quite a few tests to fail.

For the most part, getting the tests to pass required improving the
accuracy of the opcode metadata.