fix(tests): Support else blocks in glimmer tests
Previously, the test suite skipped Glimmer-style templates for any tests
that yielded to `else` or `inverse`.

In general, the test suite uses a matrix test style to describe
templates and invocations abstractly (a "blueprint"), and then compiles
the blueprint into each of the styles.

It's possible to compile `else` blocks in Glimmer:

```hbs
<TestComponent>
  <:default as |block params|>

  </:default>
  <:else>

  </:else>
</TestComponent>
```

This is not purely pedantic: the current test suite fails to test that
`yield to="inverse"` actually yields to `<:else>` blocks and this fix
therefore improves coverage of the public API.