refactor(tests): Separate invoke & template styles
The test harness previously didn't provide a way to test attributes
passed to classic components.

This commit makes it possible to use the Glimmer invocation style, which
supports attributes, alongside templates implemented using classic
features.

Example:

```ts
@test({ invokeAs: 'glimmer' })
'with ariaRole specified as an outer binding'() {
  this.render.template(
    {
      layout: 'Here!',
      attributes: {
        id: '"aria-test"',
        class: '"foo"',
        role: 'this.ariaRole'
      },
    },
    { ariaRole: 'main' }
  );

  this.assertComponent('Here!', {
    id: 'aria-test',
    class: classes('ember-view foo'),
    role: 'main',
  });
  this.assertStableRerender();
}
```

This is a test that uses a classic-style component manager and
implementation strategy but uses a Glimmer invocation style. This allows
us to test that attributes passed to classic components are preserved.