test(error-recovery): Add error recovery suites
This commit makes it possible to take any existing test suite and add
tests for no-op error recovery.

This works by wrapping each template with `{{#-try this.handleError}}`
and adding minor infrastructure to merge the wrapping properties with
the original properties.

This commit adds error recovery testing to initial render tests, but
more tests will be added in future commits.

The implementation assumes that templates all funnel through
`this.renderTemplate` in the test delegate, but this may not be
accurate, especially in rehydration tests.