refactor(references): Bake errors in more deeply
This commit significantly overhauls the reference architecture to bake
errors in at a deeper level.

After this commit, references have an additional `error` field that can
be set to indicate that a reference is in the error state.

The internal reference API was overhauled to explicitly support error
cases, most notably via a new `readReactive` function, which returns a
`Result<T, UserException>`. If the reference is in the error state,
`readReactive` returns an `Err`.

There is also an `unwrapReactive` function, which throws an exception if
the the reference is in the error state. This replaces the `valueForRef`
function.

This commit adds new primitives and updates the terminology used in the
reactive API to reflect the error recovery changes (and also to align
with Starbeam's terminology).

- `MutableCell`: a new internal primitive that reflects a single piece
  of mutable data.
- `ReadonlyCell`: a new internal primitive that reflects a single piece
  of immutable data.
- `DeeplyConstant`: replaces the `unbound` collection of APIs. Where
  `ReadonlyCell` is shallowly immutable, `DeeplyConstant` values
  produce `DeeplyConstant` property reactives.
- `FallibleFormula`: A fallible formula is a read-only function that
  can throw errors. If the compute function throws an error, the
  `FallibleFormula` will be in an error state.
- `InfallibleFormula`: An infallible formula is a read-only function
  that cannot throw errors. If the compute function throws an error,
  the `InfallibleFormula` will be in an error state.
- `Accessor`: A read-write formula is an `InfallibleFormula` on the read
  side and also has an `update` side.

At the moment, virtually all uses of `valueForRef` have been
updated to use `unwrapReactive`. This is an acceptable (but not ideal)
transition path inside of combinators (such as `concatReference`)
because they can use `FallibleFormula` to handle the error.

The VM, on the other hand, must not use `unwrapReactive` directly (and
instead must use the previously committed `vm.deref` and `vm.unwrap`
methods to handle errors in a VM-aware way).

Ultimately, combinators should also not use `unwrapReactive`, to avoid
needing so many `try/catch`es, but that can come after the VM usages
have been updated.